+++ src/sing/utaformatixProject/common.ts
@@ -1,0 +1,6 @@
+import { State } from "@/store/type";
+
+export type SongState = Pick<
+  State,
+  "tracks" | "tpqn" | "tempos" | "timeSignatures"
+>;
