+++ src/sing/utaformatixProject/import.ts
@@ -1,13 +1,14 @@
 import { Project as UfProject } from "@sevenc-nanashi/utaformatix-ts";
 import { v4 as uuidv4 } from "uuid";
-import { DEFAULT_TPQN, createDefaultTrack } from "./domain";
-import { round } from "./utility";
-import { getDoremiFromNoteNumber } from "./viewHelper";
+import { DEFAULT_TPQN, createDefaultTrack } from "@/sing/domain";
+import { round } from "@/sing/utility";
+import { getDoremiFromNoteNumber } from "@/sing/viewHelper";
 import { NoteId } from "@/type/preload";
-import { Note, SongState, Tempo, TimeSignature, Track } from "@/store/type";
+import { Note, Tempo, TimeSignature, Track } from "@/store/type";
+import { SongState } from "./common";
 
 /** UtaformatixのプロジェクトをVoicevoxの楽譜データに変換する */
-export const ufProjectToSongState = (project: UfProject): SongState => {
+export const importUtaformatixProject = (project: UfProject): SongState => {
   const convertPosition = (
     position: number,
     sourceTpqn: number,
