+++ src/store/singing.ts
@@ -88,7 +88,7 @@ import { getWorkaroundKeyRangeAdjustment } from "@/sing/workaroundKeyRangeAdjust
 import { createLogger } from "@/domain/frontend/log";
 import { noteSchema } from "@/domain/project/schema";
 import { getOrThrow } from "@/helpers/mapHelper";
-import { ufProjectToSongState } from "@/sing/ufProjectToSongState";
+import { importUtaformatixProject } from "@/sing/utaformatixProject/import";
 
 const logger = createLogger("store/singing");
 
@@ -1695,7 +1695,7 @@ export const singingStore = createPartialStore<SingingStoreTypes>({
     action: createUILockAction(
       async ({ state, dispatch }, { project, trackIndex = 0 }) => {
         const { tempos, timeSignatures, tracks, tpqn } =
-          ufProjectToSongState(project);
+          importUtaformatixProject(project);
 
         const notes = tracks[trackIndex].notes;
